refactor: rename block files
